Initial Commit with a few other things.
This is the initial commit for this application. It is a temperature
converter, and at the moment it only converts fahrenheit to celsius and
vice versa. To run it requires 'f' or 'c' to specify which scale the
given input is for the first argument, and to give the temperature as
the second.

This is a fairly small application, and it is the result of reading
through the rust documentation, so I did not split this into multiple
commits.