Fixes input issues.
This commit fixes input issues as they pertain to the temperature value.
Prior to this the data type that was expected for the app was an _i16_
which was then converted to a f64 for the calculations. This commit
changes the expected data type to f64 so there are no issues using a
float as the input value or using what could be percieved as a integer.